pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arc-memory"
version = "0.3.0"
description = "Arc Memory - Local bi-temporal knowledge graph for code repositories"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jarrod Barnes", email = "jbarnes850@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    # Core database and graph dependencies
    "networkx>=3.0",        # Graph algorithms and data structures
    "apsw>=3.40.0",         # SQLite database wrapper with advanced features

    # API and data handling
    "requests>=2.28.0",     # HTTP requests for GitHub API
    "pydantic>=2.0.0",      # Data validation and settings management
    "gql>=3.4.0",           # GraphQL client for GitHub API
    "aiohttp>=3.8.0",       # Async HTTP client for GraphQL

    # CLI and user interface
    "typer>=0.9.0",         # Command-line interface
    "rich>=13.0.0",         # Rich text and formatting in terminal
    "tqdm>=4.65.0",         # Progress bars

    # Authentication and security
    "pyjwt>=2.6.0",         # JWT token handling for GitHub auth
    "cryptography>=43.0.0", # Required for PyJWT's RS256 algorithm
    "keyring>=23.13.1",     # Secure credential storage

    # Data formats and parsing
    "pyyaml>=6.0",          # YAML parsing for ADRs
    "markdown-it-py>=2.2.0", # Markdown parsing for ADRs

    # Compression
    "zstandard>=0.20.0",    # Database compression

    # Git integration
    "gitpython>=3.1.30",    # Git repository interaction

    # Environment and configuration
    "python-dotenv>=1.0.0", # Environment variable loading
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",  # For testing async code
    "mypy>=1.3.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.2",
    "responses>=0.23.1",  # For mocking API responses in tests
]

# Testing dependencies (subset of dev dependencies)
test = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",  # For testing async code
    "responses>=0.23.1",  # For mocking API responses in tests
]

# CLI-only dependencies (for users who only need the command-line interface)
cli = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "keyring>=23.13.1",  # For storing GitHub credentials
]

# Documentation dependencies
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings>=0.20.0",
]

# Telemetry dependencies
telemetry = [
    "posthog>=3.0.1",  # Analytics and telemetry
]

# Simulation dependencies
sim = [
    "e2b-code-interpreter>=0.5.0",  # For sandbox environments
    "smolagents>=0.0.5",            # For agent-based workflow
    "openai>=1.0.0",                # For LLM integration
    "kubernetes_asyncio>=32.3.2",   # For Kubernetes integration
]

[project.urls]
"Homepage" = "https://www.arc.computer"
"Repository" = "https://github.com/Arc-Computer/arc-memory"
"Bug Tracker" = "https://github.com/Arc-Computer/arc-memory/issues"

[project.scripts]
arc = "arc_memory.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["arc_memory"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "asyncio: mark a test as an asyncio test",
]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
